requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File Parsing
python-docx==1.1.0
pywin32>=307; sys_platform == "win32"
pypandoc==1.12
markdown==3.5.1

# Web Scraping
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
readability-lxml==0.8.1
trafilatura==1.6.3

# NLP & Analysis
langdetect==1.0.9
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
keybert==0.8.1

# LLM Clients
openai==1.3.7
anthropic==0.7.8
ollama==0.1.4

# Output Generation
reportlab==4.0.7
markdown2==2.4.10

# Utilities
loguru==0.7.2
aiofiles==23.2.1
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

